Issue #14 PropertyCollectorFacade
This commit introduces pyvisdk.facade.property_collector.PropertyCollectorFacade
This facade/PropertyCollector is destroyed on the server-side when the session is terminated.
In order for different modules who share the same Vim instance be able to use the shared facades, I added a the Vim.facades dictionary to hold them
and it is cleared on login/logout.